@@ -3,7 +3,7 @@ name = "django-minio-storage"
 description = "Django file storage using the minio python client"
 license = "MIT OR Apache-2.0"
 license-files = ["LICENSE", "LICENSE-APACHE"]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
   "django>=4.2",
   "minio>=7.2.19",
